requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
jinja2
httpx
beautifulsoup4
python-dotenv
openai
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
email-validator
python-multipart
lxml
psycopg2-binary
ollama
Pillow